Rename 'IndividualApplicationPage' to 'ApplicationViewPage'
