+++ browser-tests/pages/applications.ts
@@ -4,9 +4,9 @@ export const applications = {
   applicationList: {
     firstApplicationLink: Selector('div[class^="applicationListPage"] div[class^="table_rowWrapper"]:first-of-type a'),
   },
-  applicationDetails: {
+  applicationView: {
     firstName: Selector(
-      'div[class^="individualApplicationPage"] div[class^="labelValuePair_"]:first-of-type span:nth-of-type(2)'
+      'div[class^="applicationViewPage"] div[class^="labelValuePair_"]:first-of-type span:nth-of-type(2)'
     ).nth(0),
   },
 };
