+++ src/features/applicationView/applicationViewPage.module.scss
@@ -1,6 +1,6 @@
 @import 'spacings';
 
-.individualApplicationPage {
+.applicationViewPage {
   display: grid;
   grid-template-columns: 1fr 1fr;
   grid-auto-rows: max-content;
